Regrouped Application settings into separate sections
