+++ resources/META-INF/plugin-internal.xml
@@ -31,8 +31,26 @@
         <projectService serviceImplementation="com.intellij.idea.plugin.hybris.settings.HybrisProjectSettingsComponent"/>
         <projectService serviceImplementation="com.intellij.idea.plugin.hybris.settings.HybrisDeveloperSpecificProjectSettingsComponent"/>
 
+        <projectService serviceInterface="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.RegionService"
+                        serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.impl.DefaultRegionService"/>
+        <projectService serviceInterface="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.RegionEntityService"
+                        serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.impl.DefaultRegionEntityService"/>
+        <projectService serviceInterface="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.cache.HybrisConsoleRegionsCache"
+                        serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.cache.impl.HybrisConsoleInMemoryRegionsCache"/>
+        <projectService serviceInterface="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.RegionPersistenceService"
+                        serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.impl.DefaultJsonRegionPersistenceService"/>
+        <projectService serviceInterface="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.ConsolePersistenceService"
+                        serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.impl.DefaultConsolePersistenceService"/>
+        <projectService serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.services.impl.JsonIOUtil"/>
+        <projectService serviceImplementation="com.intellij.idea.plugin.hybris.tools.remote.console.persistence.ui.listeners.HybrisConsoleQueryPanelEventManager"/>
+
         <gotoClassContributor implementation="com.intellij.idea.plugin.hybris.gotoClass.CustomGotoClassContributor"/>
         <directoryIndexExcludePolicy implementation="com.intellij.idea.plugin.hybris.indexing.HybrisDirectoryIndexExcludePolicy"/>
+
+        <applicationConfigurable id="hybris.application.settings" parentId="root"
+                                 instance="com.intellij.idea.plugin.hybris.settings.HybrisApplicationSettingsConfigurable"/>
+        <applicationConfigurable id="hybris.project.import.application.settings" parentId="hybris.application.settings"
+                                 instance="com.intellij.idea.plugin.hybris.settings.HybrisProjectImportApplicationSettingsConfigurable"/>
     </extensions>
 
     <project-components>
