+++ src/com/intellij/idea/plugin/hybris/settings/forms/HybrisApplicationSettingsForm.form
@@ -1,0 +1,96 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.intellij.idea.plugin.hybris.settings.forms.HybrisApplicationSettingsForm">
+  <grid id="27dc6" binding="mainPanel" layout-manager="GridLayoutManager" row-count="11" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+    <margin top="0" left="0" bottom="0" right="0"/>
+    <constraints>
+      <xy x="20" y="20" width="919" height="541"/>
+    </constraints>
+    <properties/>
+    <border type="none"/>
+    <children>
+      <component id="e0198" class="javax.swing.JCheckBox" binding="defaultPlatformInReadOnly">
+        <constraints>
+          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.import.settings.import.ootb.modules.read.only.label"/>
+          <toolTipText resource-bundle="i18n/HybrisBundle" key="hybris.import.settings.import.ootb.modules.read.only.tooltip"/>
+        </properties>
+      </component>
+      <component id="ce3de" class="javax.swing.JCheckBox" binding="hideEmptyMiddleFoldersCheckBox">
+        <constraints>
+          <grid row="1" column="0" row-span="1" col-span="2" vsize-policy="3" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.view.tree.hide.empty.middle.folders"/>
+        </properties>
+      </component>
+      <component id="c48aa" class="javax.swing.JCheckBox" binding="followSymlink">
+        <constraints>
+          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <selected value="true"/>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.import.followSymlink"/>
+        </properties>
+      </component>
+      <component id="8d1e4" class="javax.swing.JCheckBox" binding="withMavenSources">
+        <constraints>
+          <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <selected value="false"/>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.maven.download.sources.folders"/>
+        </properties>
+      </component>
+      <component id="ed275" class="javax.swing.JCheckBox" binding="withMavenJavadocs">
+        <constraints>
+          <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <selected value="false"/>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.maven.download.javadocs.folders"/>
+        </properties>
+      </component>
+      <component id="7d4ec" class="javax.swing.JCheckBox" binding="withStandardProvidedSources" default-binding="true">
+        <constraints>
+          <grid row="5" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.attach.standard.sources"/>
+        </properties>
+      </component>
+      <component id="a14f2" class="javax.swing.JCheckBox" binding="scanThroughExternalModule">
+        <constraints>
+          <grid row="6" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <selected value="false"/>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.import.scannExternalModules"/>
+        </properties>
+      </component>
+      <component id="7c462" class="javax.swing.JCheckBox" binding="excludeTestSources">
+        <constraints>
+          <grid row="7" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <selected value="false"/>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.project.import.excludeTestSources"/>
+        </properties>
+      </component>
+      <component id="aeae3" class="javax.swing.JCheckBox" binding="warnIfGeneratedItemsAreOutOfDateCheckBox">
+        <constraints>
+          <grid row="8" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text resource-bundle="i18n/HybrisBundle" key="hybris.ts.items.validation.settings.enabled"/>
+        </properties>
+      </component>
+      <vspacer id="d2065">
+        <constraints>
+          <grid row="9" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
+        </constraints>
+      </vspacer>
+    </children>
+  </grid>
+</form>
