+++ src/com/intellij/idea/plugin/hybris/settings/forms/HybrisImpexApplicationSettingsForm.form
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.intellij.idea.plugin.hybris.settings.forms.HybrisImpexApplicationSettingsForm">
+  <grid id="27dc6" binding="mainPanel" layout-manager="GridLayoutManager" row-count="3" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+    <margin top="0" left="0" bottom="0" right="0"/>
+    <constraints>
+      <xy x="20" y="20" width="919" height="73"/>
+    </constraints>
+    <properties/>
+    <border type="none"/>
+    <children>
+      <component id="b4c8b" class="javax.swing.JCheckBox" binding="enableFoldingCheckBox" default-binding="true">
+        <constraints>
+          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text value="Enable folding"/>
+        </properties>
+      </component>
+      <component id="fbb4b" class="javax.swing.JCheckBox" binding="useSmartFoldingCheckBox" default-binding="true">
+        <constraints>
+          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <text value="Use smart folding"/>
+          <toolTipText value="if enabled, folded regions will constructed in a way to show as much information as possible.  "/>
+        </properties>
+      </component>
+      <vspacer id="47809">
+        <constraints>
+          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
+        </constraints>
+      </vspacer>
+    </children>
+  </grid>
+</form>
